+++ trixnity-messenger-compose-view/src/desktopMain/kotlin/de/connect2x/messenger/desktop/Notifications.kt
@@ -2,13 +2,16 @@ package de.connect2x.messenger.desktop
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.window.Notification
-import androidx.compose.ui.window.TrayState
+import androidx.compose.ui.graphics.toAwtImage
 import de.connect2x.messenger.compose.view.DI
 import de.connect2x.messenger.compose.view.IsFocused
 import de.connect2x.messenger.compose.view.files.toImageBitmap
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
+import de.connect2x.sysnotify.Notification
+import de.connect2x.sysnotify.NotificationHandler
+import de.connect2x.sysnotify.NotificationIcon
+import de.connect2x.sysnotify.fromBufferedImage
 import de.connect2x.trixnity.messenger.MatrixClients
 import de.connect2x.trixnity.messenger.MatrixMessenger
 import de.connect2x.trixnity.messenger.MatrixMessengerAccountSettings
@@ -39,14 +42,13 @@ import net.folivo.trixnity.core.model.events.m.room.RoomMessageEventContent
 import net.folivo.trixnity.core.model.events.roomIdOrNull
 import net.folivo.trixnity.core.model.events.senderOrNull
 import net.folivo.trixnity.core.model.push.PushAction
-import javax.sound.sampled.AudioSystem
 
 private val log = KotlinLogging.logger { }
 
 @Composable
 fun Notifications(
     matrixMessenger: MatrixMessenger,
-    trayState: TrayState,
+    notificationHandler: NotificationHandler
 ) {
     val i18n = DI.get<I18nView>()
     val maxAvatarSize = DI.get<MatrixMessengerConfiguration>().avatarMaxSize
@@ -55,7 +57,14 @@ fun Notifications(
     LaunchedEffect(windowIsFocused) {
         withContext(Dispatchers.Default) {
             val roomNameComputation = matrixMessenger.di.get<RoomName>()
-            whenSyncIsRunning(matrixMessenger, windowIsFocused, roomNameComputation, trayState, i18n, maxAvatarSize)
+            whenSyncIsRunning(
+                matrixMessenger,
+                windowIsFocused,
+                roomNameComputation,
+                i18n,
+                maxAvatarSize,
+                notificationHandler
+            )
         }
     }
 }
@@ -64,9 +73,9 @@ private suspend fun whenSyncIsRunning(
     matrixMessenger: MatrixMessenger,
     windowIsFocused: Boolean,
     roomNameComputation: RoomName,
-    trayState: TrayState,
     i18n: I18nView,
-    maxAvatarSize: Long
+    maxAvatarSize: Long,
+    notificationHandler: NotificationHandler
 ) {
     val settings = matrixMessenger.di.get<MatrixMessengerSettingsHolder>()
     matrixMessenger.di.get<MatrixClients>().scopedCollectLatest { matrixClients ->
@@ -81,13 +90,7 @@ private suspend fun whenSyncIsRunning(
                         if (currentSettings.platformNotifications.notificationsPlaySound &&
                             notification.actions.any { it is PushAction.SetSoundTweak }
                         ) {
-                            withContext(Dispatchers.IO) {
-                                MessengerTrayIcon::class.java.getResourceAsStream("/ding.wav")
-                                    ?.buffered()
-                                    ?.let(AudioSystem::getAudioInputStream)
-                                    ?.let { AudioSystem.getClip().apply { open(it) } }
-                                    ?.start()
-                            }
+                            // TODO: implement a sound flag in SysNotify 1.6.2
                         }
                         if (currentSettings.platformNotifications.notificationsShowPopup) {
                             val room = notification.event.roomIdOrNull?.let { matrixClient.room.getById(it).first() }
@@ -102,8 +105,9 @@ private suspend fun whenSyncIsRunning(
                                 isDirect,
                                 roomName,
                                 i18n,
-                                maxAvatarSize
-                            )?.let { trayState.sendNotification(it) }
+                                maxAvatarSize,
+                                notificationHandler
+                            )
                         }
                     }
                 }
@@ -120,8 +124,9 @@ private suspend fun displayNotification(
     isDirect: Boolean,
     roomName: String,
     i18n: I18nView,
-    maxAvatarSize: Long
-): Notification? {
+    maxAvatarSize: Long,
+    notificationHandler: NotificationHandler
+) {
     event.roomIdOrNull?.let { roomId ->
         val message = when {
             currentSettings.platformNotifications.notificationsShowText.not() -> "(${i18n.newMessage()})"
@@ -131,7 +136,7 @@ private suspend fun displayNotification(
         }
 
         if (message != null) {
-            val (username, _) = event.senderOrNull?.let { sender ->
+            val (username, image) = event.senderOrNull?.let { sender ->
                 val user = matrixClient.user.getById(roomId, sender).first()
                 val image = user?.avatarUrl?.let { avatarUrl ->
                     matrixClient.media.getThumbnail(avatarUrl, avatarSize().toLong(), avatarSize().toLong())
@@ -141,13 +146,15 @@ private suspend fun displayNotification(
                 user?.name to image
             } ?: (null to null)
 
-            val title = if (isDirect) username.orEmpty() else roomName
-            val text = if (isDirect) message else "$username: $message"
-
             log.debug { "notification will appear" }
-
-            return Notification(title, text)
+            notificationHandler.push(
+                Notification(
+                    title = if (isDirect) username.orEmpty() else roomName,
+                    group = "${matrixClient.userId}-$roomId",
+                    description = if (isDirect) message else "$username: $message",
+                    icon = image?.toAwtImage()?.let(NotificationIcon::fromBufferedImage)
+                )
+            )
         }
     } ?: log.warn { "cannot find roomId for event ${event.idOrNull}" }
-    return null
 }