+++ trixnity-messenger/src/iosMain/kotlin/de/connect2x/trixnity/messenger/CreateRepositoriesModule.ios.kt
@@ -1,10 +1,41 @@
 package de.connect2x.trixnity.messenger
 
+import androidx.room.Room
+import androidx.room.RoomDatabase
+import androidx.sqlite.driver.bundled.BundledSQLiteDriver
+import de.connect2x.trixnity.messenger.util.RootPath
+import de.connect2x.trixnity.messenger.util.SecretByteArray
+import net.folivo.trixnity.client.store.repository.room.TrixnityRoomDatabase
+import net.folivo.trixnity.client.store.repository.room.createRoomRepositoriesModule
+import net.folivo.trixnity.core.model.UserId
 import org.koin.core.module.Module
 import org.koin.dsl.module
 
-actual fun platformEncryptRepositoryModule(): Module = module {
-    single<EncryptRepository> {
-        EncryptRepository { false }
+actual fun platformCreateRepositoriesModuleModule(): Module = module {
+    single<CreateRepositoriesModule> {
+        val rootPath = get<RootPath>()
+
+        object : CreateRepositoriesModule {
+            override suspend fun create(userId: UserId): CreateRepositoriesModule.CreateResult {
+                return CreateRepositoriesModule.CreateResult(
+                    module = createRoomRepositoriesModule(db(userId)),
+                    databaseKey = null,
+                )
+            }
+
+            override suspend fun load(
+                userId: UserId,
+                databaseKey: SecretByteArray?,
+            ): Module {
+                return createRoomRepositoriesModule(db(userId))
+            }
+
+            private fun db(userId: UserId): RoomDatabase.Builder<TrixnityRoomDatabase> =
+                Room.databaseBuilder<TrixnityRoomDatabase>(
+                    rootPath.forAccountDatabase(userId).resolve("database").toString()
+                ).apply {
+                    setDriver(BundledSQLiteDriver())
+                }
+        }
     }
 }